fix: change Stack to Skill, add new Skill List
