+++ my-profile-card/src/components/Skill.js
@@ -1,12 +1,12 @@
 import * as IoIcons from "react-icons/io5";
 
-export default function Stack(props) {
+export default function Skill(props) {
   // Get the icon component dynamically
   const IconComponent = IoIcons[props.icon];
 
   return (
     <div
-      className="stack"
+      className="skill"
       key={props.name}
       style={{ backgroundColor: props.bgColor }}
     >
